Revert "Merge branch 'SIW-1466-add-login-io-backend'"
This reverts commit 2926d8422e6aa2509ca54ec64b776992ffd7f1fc, reversing
changes made to d1de83816744880b4ba85c425a15d135d794b429.
