+++ .gitignore
@@ -77,6 +77,3 @@ example/.env
 
 #Generated by openapi
 src/client/generated/*.ts
-
-#Android bundle
-example/android/app/src/main/assets/index.android.bundle
