+++ example/.env.example
@@ -1,8 +1,9 @@
 WALLET_PROVIDER_BASE_URL='example' # Base URL for the Wallet provider
-WALLET_PID_PROVIDER_BASE_URL='example' # Base URL for the PID provider PRE
-WALLET_EAA_PROVIDER_BASE_URL='example' # Base URL for the EAA provider PRE
-IPZS_AUTH_TOKEN='example' # Auth token for IPZS
+WALLET_PROVIDER_AUTH_TOKEN='example' # Bearer Token for the Wallet provider mediated via io-backend
+WALLET_PID_PROVIDER_BASE_URL='example' # Base URL for the PID provider
+WALLET_EAA_PROVIDER_BASE_URL='example' # Base URL for the EAA provider
 REDIRECT_URI='example' # Redirect URI used to detect a successful authentication. On Android it must be registered in the AndroidManifest.xml
 GOOGLE_CLOUD_PROJECT_NUMBER='1234567890' # Google Cloud Project Number for the Google Play Integrity API
+CIE_PIN='123456' # Pin of CIE to use with L3 login
 CIE_UAT='true' #If you want to use CIE with uat environment or prod
 SPID_IDPHINT='https://demo.spid.gov.it' # idphint to use with SPID
