+++ example/package.json
@@ -16,17 +16,10 @@
     "@pagopa/io-react-native-jwt": "^1.2.0",
     "@pagopa/io-react-native-login-utils": "^1.0.1",
     "@pagopa/react-native-cie": "^1.3.0",
-    "@react-native-async-storage/async-storage": "^1.24.0",
-    "@reduxjs/toolkit": "^2.2.7",
     "react": "18.2.0",
     "react-native": "0.72.14",
     "react-native-safe-area-context": "^4.9.0",
-    "react-native-webview": "^13.10.5",
-    "react-redux": "^9.1.2",
-    "redux": "^5.0.1",
-    "redux-persist": "^6.0.0",
-    "typesafe-actions": "^5.1.0",
-    "url-parse": "^1.5.10"
+    "react-native-webview": "^13.10.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
@@ -34,7 +27,6 @@
     "@babel/runtime": "^7.20.0",
     "@react-native/eslint-config": "^0.72.2",
     "@react-native/metro-config": "^0.72.12",
-    "@types/url-parse": "^1.4.11",
     "babel-plugin-module-resolver": "^5.0.0",
     "metro-react-native-babel-preset": "^0.76.9",
     "react-native-dotenv": "^3.4.11",
