+++ example/src/config.ts
@@ -1,10 +1,11 @@
 declare module "@env" {
   export const WALLET_PROVIDER_BASE_URL: string;
+  export const WALLET_PROVIDER_AUTH_TOKEN: string;
   export const WALLET_PID_PROVIDER_BASE_URL: string;
   export const WALLET_EAA_PROVIDER_BASE_URL: string;
-  export const IPZS_AUTH_TOKEN: string;
   export const REDIRECT_URI: string;
   export const GOOGLE_CLOUD_PROJECT_NUMBER: string;
+  export const CIE_PIN: string;
   export const CIE_UAT: string;
   export const SPID_IDPHINT: string;
 }
