--- example/src/store/dispatch.ts
@@ -1,12 +1,0 @@
-import { useDispatch } from "react-redux";
-import type { store } from "./store";
-
-/**
- * Type definition for the dispatch function of the Redux store.
- */
-type AppDispatch = typeof store.dispatch;
-
-/**
- * Hook to use the Redux dispatch function with the correct type.
- */
-export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
