--- example/src/store/store.ts
@@ -1,36 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import { sessionReducer } from "./reducers/sesssion";
-import {
-  FLUSH,
-  PAUSE,
-  PERSIST,
-  persistStore,
-  PURGE,
-  REGISTER,
-  REHYDRATE,
-} from "redux-persist";
-
-/**
- * Redux store configuration.
- */
-export const store = configureStore({
-  reducer: {
-    session: sessionReducer,
-  },
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware({
-      serializableCheck: {
-        ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER], // Ignore all the action types dispatched by Redux Persist
-      },
-    }),
-});
-
-/**
- * Redux persistor configuration used in the root component with {@link PersistGate}.
- */
-export const persistor = persistStore(store);
-
-/**
- * Type definition for the root state of the Redux store.
- */
-export type RootState = ReturnType<typeof store.getState>;
