+++ example/src/utils/fetch.ts
@@ -1,11 +1,4 @@
-import {
-  IPZS_AUTH_TOKEN,
-  WALLET_EAA_PROVIDER_BASE_URL,
-  WALLET_PID_PROVIDER_BASE_URL,
-  WALLET_PROVIDER_BASE_URL,
-} from "@env";
-import { selectIoAuthToken } from "../store/reducers/sesssion";
-import { store } from "../store/store";
+import { WALLET_PROVIDER_AUTH_TOKEN, WALLET_PROVIDER_BASE_URL } from "@env";
 
 interface AuthHeaders {
   Authorization?: string;
@@ -22,33 +15,14 @@ function addAuthHeaders(options: RequestInit, authHeaders: AuthHeaders) {
 }
 
 export default function appFetch(request: RequestInfo, options: RequestInit) {
-  const state = store.getState();
-  const authToken = selectIoAuthToken(state);
   const requestUrl =
     typeof request === "string" ? new URL(request) : new URL(request.url);
-
-  const authHeaders: AuthHeaders = (function () {
-    switch (requestUrl.origin) {
-      case new URL(WALLET_PROVIDER_BASE_URL).origin: {
-        return {
-          Authorization: `Bearer ${authToken}`,
-        };
-      }
-      case new URL(WALLET_PID_PROVIDER_BASE_URL).origin: {
-        return {
-          AuthorizationIPZS: `${IPZS_AUTH_TOKEN}`,
-        };
-      }
-      case new URL(WALLET_EAA_PROVIDER_BASE_URL).origin: {
-        return {
-          AuthorizationIPZS: `${IPZS_AUTH_TOKEN}`,
-        };
-      }
-      default: {
-        return {};
-      }
-    }
-  })();
+  const authHeaders: AuthHeaders =
+    requestUrl.origin === new URL(WALLET_PROVIDER_BASE_URL).origin
+      ? {
+          Authorization: `Bearer ${WALLET_PROVIDER_AUTH_TOKEN}`,
+        }
+      : {};
 
   return fetch(request, addAuthHeaders(options, authHeaders));
 }
