--- example/src/utils/idps.ts
@@ -1,74 +1,0 @@
-export const idps = [
-  {
-    id: "arubaid",
-    name: "Aruba ID",
-    logo: "",
-    profileUrl: "https://selfcarespid.aruba.it",
-  },
-  {
-    id: "ehtid",
-    name: "Etna ID",
-    logo: "",
-    profileUrl: "https://etnaid.eht.eu/",
-  },
-  {
-    id: "infocamereid",
-    name: "ID InfoCamere",
-    logo: "",
-    profileUrl: "https://selfcarespid.infocamere.it/spid-selfCare/#/login",
-  },
-  {
-    id: "infocertid",
-    name: "InfoCert ID",
-    logo: "",
-    profileUrl: "https://my.infocert.it/selfcare",
-  },
-  {
-    id: "intesiid",
-    name: "Intesi Group SPID",
-    logo: "",
-    profileUrl: "https://spid.intesigroup.com",
-  },
-  {
-    id: "lepidaid",
-    name: "Lepida ID",
-    logo: "",
-    profileUrl: "https://id.lepida.it/",
-  },
-  {
-    id: "namirialid",
-    name: "Namirial ID",
-    logo: "",
-    profileUrl: "https://idp.namirialtsp.com/idp",
-  },
-  {
-    id: "posteid",
-    name: "Poste ID",
-    logo: "",
-    profileUrl: "https://posteid.poste.it/private/cruscotto.shtml",
-  },
-  {
-    id: "sielteid",
-    name: "Sielte ID",
-    logo: "",
-    profileUrl: "https://myid.sieltecloud.it/profile/",
-  },
-  {
-    id: "spiditalia",
-    name: "SpidItalia",
-    logo: "",
-    profileUrl: "https://spid.register.it",
-  },
-  {
-    id: "timid",
-    name: "TIM id",
-    logo: "",
-    profileUrl: "https://id.tim.it/identity/private/",
-  },
-  {
-    id: "teamsystemid",
-    name: "TeamSystem ID",
-    logo: "",
-    profileUrl: "https://identity.teamsystem.com/",
-  },
-];
