+++ src/utils/par.ts
@@ -25,10 +25,10 @@ export const AuthorizationDetails = z.array(AuthorizationDetail);
 export const makeParRequest =
   ({
     wiaCryptoContext,
-    appFetch,
+    appFetch = fetch,
   }: {
     wiaCryptoContext: CryptoContext;
-    appFetch: GlobalFetch["fetch"];
+    appFetch?: GlobalFetch["fetch"];
   }) =>
   async (
     clientId: string,
